+++ Classes/CrawlStrategy/GuzzleExecutionStrategy.php
@@ -40,7 +40,7 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategyInte
     /**
      * Sets up a CURL / Guzzle Request for fetching the request.
      *
-     * @return bool|mixed
+     * @return array|false See CrawlStrategyInterface::fetchUrlContents()
      */
     public function fetchUrlContents(UriInterface $url, string $crawlerId)
     {
@@ -56,7 +56,15 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategyInte
         try {
             $url = (string) $url;
             $response = $this->getResponse($url, $options);
-            return unserialize($response->getHeaderLine('X-T3Crawler-Meta'));
+            if ($response->hasHeader('X-T3Crawler-Meta')) {
+                return unserialize($response->getHeaderLine('X-T3Crawler-Meta'));
+            }
+            return [
+                'errorlog' => ['Response has no X-T3Crawler-Meta header'],
+                'vars' => [
+                    'status' => $response->getStatusCode() . ' ' . $response->getReasonPhrase(),
+                ],
+            ];
         } catch (RequestException $e) {
             $response = $e->getResponse();
             $message = ($response ? $response->getStatusCode() : 0)
@@ -69,7 +77,9 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategyInte
                     'crawlerId' => $crawlerId,
                 ]
             );
-            return $message;
+            return [
+                'errorlog' => [$message],
+            ];
         } catch (ConnectException $e) {
             $message = $e->getCode() . chr(32) . $e->getMessage();
 
@@ -79,7 +89,9 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategyInte
                     'crawlerId' => $crawlerId,
                 ]
             );
-            return $message;
+            return [
+                'errorlog' => [$message],
+            ];
         }
     }