+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Added
 * Extract error information in SubProcess crawl strategy [@cweiske](https://github.com/cweiske)
+* Expose error when crawl response has not X-T3Crawler-Meta header [@cweiske](https://github.com/cweiske)
 
 ### Changed
 
 ### Fixed
 * Set guzzle connect timeout to 5 seconds [@cweiske](https://github.com/cweiske)
+* Prevent warning when guzzle crawl requests return errors [@cweiske](https://github.com/cweiske)
 
 ### Deprecated
 #### Classes